% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Creates a map of accident locations for one as specified by the year parameter for the 
state specified by FIPS code using the state.num parameter.}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{FIPS code of the state to create a map for}

\item{year}{Year of FARS input data to us}
}
\value{
A map of accident locations for state with FIPS code tate.num in year year.
}
\description{
This function requires the package map to be loaded or if map is installed to run the assignment
stateMapEnv <- "R_MAP_DATA_DIR".  I don't know how to properly include this warning in the documentation.
}
\examples{
\dontrun{
fars_map_state(24,2013)
}

}
