% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read FARS data}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{Name of csv file with FARS data}
}
\value{
This function returns a dataframe with FARS data if the file exists
}
\description{
Creates a dataframe from csv formatted file of data from the US National Highway Traffic Safety Administration's 
 Fatality Analysis #'  Reporting System (FARS), which is a nationwide census providing the American public yearly 
 data regarding fatal injuries suffered in motor vehicle traffic crashes.  If full file path is not provided, this
 function will look in the current working directory for the file
}
\examples{
\dontrun{
fars_data <- fars_read(accident_2013.csv.bz2)
}

}
