% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Creates a dataframe from FARS data from multiple years as specified by the years parameter}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{Vector of Years of FARS input data to read}
}
\value{
A FARS dataframe for the specified year
}
\description{
This function also requires the %>% function from the package dplyr, so you that package must be loaded 
for the function to work.  (It seems like there should be a tag to indicate things like this, ditto with map below)
}
\examples{
\dontrun{
fars_data <- fars_read_years(c(2013, 2014, 2015)
}

}
