% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Creates a dataframe summarizing the number of observations by in FARS data from multiple years as specified by the 
years parameter}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{Vector of Years of FARS input data to read}
}
\value{
A dataframe summarizing the number of FARS data observations by year for the specified years
}
\description{
Creates a dataframe summarizing the number of observations by in FARS data from multiple years as specified by the 
years parameter
}
\examples{
\dontrun{
fars_summary <- fars_summarize_years(c(2013, 2014, 2015)
}

}
