% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Make a FARS csv input data filename given a year}
\usage{
make_filename(year)
}
\arguments{
\item{year}{Year of FARS input data to read}
}
\value{
A FARS data filename for a specified year
}
\description{
Make a FARS csv input data filename given a year
}
\examples{
\dontrun{
fars_fileame <- make_filename(2013)
}

}
